add sonner and refactor delete cowork
