+++ apps/front/package.json
@@ -26,6 +26,7 @@
     "react-icons": "^4.8.0",
     "react-query": "^3.39.3",
     "react-table": "^7.8.0",
+    "sonner": "^0.3.5",
     "ui": "*",
     "zod": "^3.21.4"
   },
